Remove 84 code smells

+++ module-cinematic/src/main/java/fr/perrier/dungeons/module/cinematic/execution/CinematicPlayer.java
@@ -165,7 +165,7 @@ public class CinematicPlayer {
         } else if (before != null) {
             // Between two waypoints — interpolate
             int segmentLength = after.getTick() - before.getTick();
-            double progress = segmentLength > 0 ? (double)(currentTick - before.getTick()) / segmentLength : 0;
+            double progress = segmentLength > 0 ? (currentTick - before.getTick()) / segmentLength : 0;
 
             double[] pos = CameraInterpolation.interpolate(before, after, progress, before.getInterpolation());
             callback.setCameraPosition(pos[0], pos[1], pos[2], (float) pos[3], (float) pos[4]);
@@ -196,7 +196,7 @@ public class CinematicPlayer {
                     before.getYaw(), before.getPitch(), before.getAnimation());
         } else if (before != null) {
             int segmentLength = after.getTick() - before.getTick();
-            double progress = segmentLength > 0 ? (double)(currentTick - before.getTick()) / segmentLength : 0;
+            double progress = segmentLength > 0 ? (currentTick - before.getTick()) / segmentLength : 0;
             double x = before.getX() + (after.getX() - before.getX()) * progress;
             double y = before.getY() + (after.getY() - before.getY()) * progress;
             double z = before.getZ() + (after.getZ() - before.getZ()) * progress;
